+++ tomcat/src/main/java/nextstep/web/HelloController.java
@@ -1,6 +1,6 @@
 package nextstep.web;
 
-import org.apache.coyote.http11.mvc.AbstractController;
+import org.apache.coyote.http11.controller.AbstractController;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
