+++ tomcat/src/main/java/nextstep/web/RegisterController.java
@@ -1,7 +1,7 @@
 package nextstep.web;
 
 import nextstep.jwp.application.RegisterService;
-import org.apache.coyote.http11.mvc.AbstractController;
+import org.apache.coyote.http11.controller.AbstractController;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
