+++ tomcat/src/main/java/org/apache/coyote/http11/controller/Controller.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.mvc;
+package org.apache.coyote.http11.controller;
 
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
