+++ tomcat/src/main/java/org/apache/coyote/http11/controller/FrontController.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.mvc;
+package org.apache.coyote.http11.controller;
 
 import java.io.IOException;
 import org.apache.coyote.http11.request.HttpRequest;
