+++ tomcat/src/main/java/org/apache/coyote/http11/view/StaticResourceView.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.mvc.view;
+package org.apache.coyote.http11.view;
 
 import org.apache.coyote.http11.common.ResourceContentType;
