+++ tomcat/src/test/java/org/apache/coyote/http11/mvc/view/StaticResourceViewTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.io.IOException;
+import org.apache.coyote.http11.view.StaticResourceView;
+import org.apache.coyote.http11.view.View;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
